{"schema":"decodex/commit/1","summary":"Split MCP tools registry modules","authority":"XY-1115"}
