+++ apps/decodex/src/mcp/tools/autonomy.rs
@@ -1,0 +1,60 @@
+use crate::mcp::{self, McpCapabilityProfile, McpTool, tool_schemas, tools::entry};
+
+pub(super) fn mcp_autonomy_tools() -> Vec<McpTool> {
+	vec![
+		entry::mcp_tool_entry(
+			McpCapabilityProfile::Plan,
+			mcp::TOOL_AUTONOMY_DRAFT_OBJECTIVE,
+			"Decodex Autonomy Draft Objective",
+			"Validate or persist a draft Objective Contract without granting acceptance authority.",
+			tool_schemas::autonomy_draft_objective_tool_input_schema(),
+			tool_schemas::autonomy_objective_tool_output_schema(),
+			false,
+		),
+		entry::mcp_tool_entry(
+			McpCapabilityProfile::Plan,
+			mcp::TOOL_AUTONOMY_ACCEPT_OBJECTIVE,
+			"Decodex Autonomy Accept Objective",
+			"Accept a draft Objective Contract version as project-level autonomy authority without starting execution.",
+			tool_schemas::autonomy_accept_objective_tool_input_schema(),
+			tool_schemas::autonomy_objective_tool_output_schema(),
+			false,
+		),
+		entry::mcp_tool_entry(
+			McpCapabilityProfile::Plan,
+			mcp::TOOL_AUTONOMY_SUBMIT_SIGNAL,
+			"Decodex Autonomy Submit Signal",
+			"Validate or persist proposal-only autonomy signal evidence under an accepted objective.",
+			tool_schemas::autonomy_submit_signal_tool_input_schema(),
+			tool_schemas::autonomy_signal_tool_output_schema(),
+			false,
+		),
+		entry::mcp_tool_entry(
+			McpCapabilityProfile::Plan,
+			mcp::TOOL_AUTONOMY_COMPILE_PROPOSAL,
+			"Decodex Autonomy Compile Proposal",
+			"Compile or persist non-executable autonomy proposal evidence from accepted objective-bound signals.",
+			tool_schemas::autonomy_compile_proposal_tool_input_schema(),
+			tool_schemas::autonomy_proposal_tool_output_schema(),
+			false,
+		),
+		entry::mcp_tool_entry(
+			McpCapabilityProfile::Plan,
+			mcp::TOOL_AUTONOMY_CHALLENGE_PROPOSAL,
+			"Decodex Autonomy Challenge Proposal",
+			"Dry-run or record challenge evidence for an autonomy proposal without making it acceptance authority.",
+			tool_schemas::autonomy_challenge_proposal_tool_input_schema(),
+			tool_schemas::autonomy_challenge_tool_output_schema(),
+			false,
+		),
+		entry::mcp_tool_entry(
+			McpCapabilityProfile::Plan,
+			mcp::TOOL_AUTONOMY_REQUEST_PROMOTION,
+			"Decodex Autonomy Request Promotion",
+			"Inspect or explicitly accept an autonomy proposal into a latent Decision Contract candidate.",
+			tool_schemas::autonomy_request_promotion_tool_input_schema(),
+			tool_schemas::autonomy_promotion_request_tool_output_schema(),
+			false,
+		),
+	]
+}
