+++ apps/decodex/src/mcp/tools/profiles.rs
@@ -1,0 +1,25 @@
+use crate::mcp::{
+	McpCapabilityProfile, TOOL_AUTONOMY_ACCEPT_OBJECTIVE, TOOL_AUTONOMY_CHALLENGE_PROPOSAL,
+	TOOL_AUTONOMY_COMPILE_PROPOSAL, TOOL_AUTONOMY_DRAFT_OBJECTIVE, TOOL_AUTONOMY_REQUEST_PROMOTION,
+	TOOL_AUTONOMY_SUBMIT_SIGNAL, TOOL_INTAKE_GOAL, TOOL_LANE_CONTROL, TOOL_OBSERVE, TOOL_PLAN,
+	TOOL_PROJECT_CONTROL, TOOL_RESEARCH_COMPILE, TOOL_RESEARCH_PROMOTE,
+};
+
+pub(in crate::mcp) fn tool_required_profile(name: &str) -> Option<McpCapabilityProfile> {
+	match name {
+		TOOL_OBSERVE => Some(McpCapabilityProfile::Observe),
+		TOOL_PLAN => Some(McpCapabilityProfile::Plan),
+		TOOL_RESEARCH_COMPILE
+		| TOOL_RESEARCH_PROMOTE
+		| TOOL_INTAKE_GOAL
+		| TOOL_AUTONOMY_DRAFT_OBJECTIVE
+		| TOOL_AUTONOMY_ACCEPT_OBJECTIVE
+		| TOOL_AUTONOMY_SUBMIT_SIGNAL
+		| TOOL_AUTONOMY_COMPILE_PROPOSAL
+		| TOOL_AUTONOMY_CHALLENGE_PROPOSAL
+		| TOOL_AUTONOMY_REQUEST_PROMOTION => Some(McpCapabilityProfile::Plan),
+		TOOL_LANE_CONTROL => Some(McpCapabilityProfile::Operate),
+		TOOL_PROJECT_CONTROL => Some(McpCapabilityProfile::Admin),
+		_ => None,
+	}
+}
